+++ lib/fakturoid/client/expense.rb
@@ -14,9 +14,13 @@ module Fakturoid
         get_request("expenses/#{id}.json")
       end
 
-      def self.search(query)
+      def self.search(query, params = {})
         validate_search_query(query)
-        get_request("expenses/search.json", request_params: { query: query })
+
+        request_params = permit_params(params, :page)
+        request_params[:query] = query
+
+        get_request("expenses/search.json", request_params: request_params)
       end
 
       def self.fire(id, event, params = {})